requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
fuzzywuzzy==0.18.0
pycparser==2.20
pydo.txt==1.1
python-Levenshtein==0.12.2
python-telegram-bot==13.2
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
